# backend/tests/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "NotificacionResponse": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "leida": {
            "title": "Leida",
            "type": "boolean"
          },
          "mensaje": {
            "title": "Mensaje",
            "type": "string"
          },
          "reserva_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reserva Id"
          },
          "tipo": {
            "$ref": "#/components/schemas/TipoNotificacion"
          },
          "usuario_id": {
            "title": "Usuario Id",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "usuario_id",
          "reserva_id",
          "tipo",
          "leida",
          "created_at",
          "mensaje"
        ],
        "title": "NotificacionResponse",
        "type": "object"
      },
      "NotificacionesSinLeerResponse": {
        "properties": {
          "cantidad": {
            "title": "Cantidad",
            "type": "integer"
          }
        },
        "required": [
          "cantidad"
        ],
        "title": "NotificacionesSinLeerResponse",
        "type": "object"
      },
...
      "TipoNotificacion": {
        "description": "Tipos de notificación internos.\n\nOrigen verificado de los valores: el constraint `notificaciones_tipo_check`\nexiste en base de datos — declarado en `app/models/notificacion.py`\n(`__table_args__`) y aplicado por `app/migrations.py` (DROP IF EXISTS +\nADD CHECK `tipo IN ('Pendiente', 'Aprobada', 'Rechazada', 'Cancelada', 'Actualizada')`).\nSe conservan exactamente esos valores.",
        "enum": [
          "Pendiente",
          "Aprobada",
          "Rechazada",
          "Cancelada",
          "Actualizada"
        ],
        "title": "TipoNotificacion",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "cookieAuth": {
        "in": "cookie",
        "name": "access_token",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "API para gestionar recursos y sus reservas",
    "title": "Gestión de Reservas de Recursos Institucionales",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/notificaciones": {
      "get": {
        "operationId": "listar_mis_notificaciones_notificaciones_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/NotificacionResponse"
                  },
                  "title": "Response Listar Mis Notificaciones Notificaciones Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Listar Mis Notificaciones",
        "tags": [
          "notificaciones"
        ]
      }
    },
    "/notificaciones/leer-todas": {
      "patch": {
        "operationId": "marcar_todas_como_leidas_notificaciones_leer_todas_patch",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificacionesSinLeerResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Marcar Todas Como Leidas",
        "tags": [
          "notificaciones"
        ]
      }
    },
    "/notificaciones/sin-leer/count": {
      "get": {
        "operationId": "contar_mis_notificaciones_sin_leer_notificaciones_sin_leer_count_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificacionesSinLeerResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Contar Mis Notificaciones Sin Leer",
        "tags": [
          "notificaciones"
        ]
      }
    },
    "/notificaciones/{notificacion_id}/leer": {
      "patch": {
        "operationId": "marcar_como_leida_notificaciones__notificacion_id__leer_patch",
        "parameters": [
          {
            "in": "path",
            "name": "notificacion_id",
            "required": true,
            "schema": {
              "title": "Notificacion Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificacionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Marcar Como Leida",
        "tags": [
          "notificaciones"
        ]
      }
    },
...
  }
}